Initial Project
My initial commit is done by the git Desktop, when I clone an empty project from github using Desktop to my local projects/repos/github.   This contains only a  UNITY .gitignore, a License, and a readme.

This "Initial Project" commit is the result of creating a new project in unity, setting up it's project settings for version control, renaming the default scene and Scenes Folder to match my naming schemes.  And nothing else.