Roughing in the shape of things.
Added a LevelManager, which places items on the counter using an "ItemPlacer" script, which is implemented with a Generic Queue.  After the Item is SetActive, it's "PriceScript" sets the item's price (simple assignment at this point--I have yet to hook up the cLevel, cLevelManager, cCurrencyValue and cRange classes, which will enable the LevelManager to assign level-specific, random amounts.