Now taking items off the counter and putting items on the counter.
I learned a really neat (although obvious) trick, I have my ItemTrigger object, which waits for items to bump into it, and I placed a public reference to my LevelManager (a.k.a. GameManager) which contains the "ItemPlacer" script, so the ItemTrigger object can initiate the queueing of items and placement of items off the queue via the public methods in the ItemPlacer script.