Converted cLevel so that it is completely static.
cItemPlacer now is keeping track of the number of items placed, and when ITEMSPERLEVEL (10) is reached, it calls the static cLevel.LevelUp(), showing that the levels are available to the public.

Next, I want to hook up cCurrencyValue, so that I am using that to develop Random prices for the items.